Add Headless-Setup.sh
The current setup script (Sleepy-Pi-Setup.sh) doesn't work if you
attempt to run it on a headless system. This is due to the commands
that install the Arduino IDE, as it depends on a desktop environment.

Headless-Setup.sh is a copy of Sleepy-Pi-Setup.sh with the offending
sections removed. It was created as a separate file for devolopment
purposes but is intended to merge with Sleepy-Pi-Setup.sh at a later
date.